webserver/requirements.txt
Pillow
flask
uuid==1.30
tensorflow-cpu==2.11.0
numpy
python-dotenv
h5py==3.7.0
scipy==1.10.1
paho-mqtt==1.6.1
